Separate lock.h: macOS
