+++ include/os/macos/etcpal/os_signal.h
@@ -1,0 +1,53 @@
+/******************************************************************************
+ * Copyright 2020 ETC Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *******************************************************************************
+ * This file is a part of EtcPal. For more information, go to:
+ * https://github.com/ETCLabs/EtcPal
+ ******************************************************************************/
+
+#ifndef ETCPAL_OS_SIGNAL_H_
+#define ETCPAL_OS_SIGNAL_H_
+
+#include <stdbool.h>
+#include <pthread.h>
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+typedef struct
+{
+  bool            valid;
+  bool            signaled;
+  pthread_cond_t  cond;
+  pthread_mutex_t mutex;
+} etcpal_signal_t;
+
+#define ETCPAL_SIGNAL_HAS_TIMED_WAIT 0
+#define ETCPAL_SIGNAL_HAS_POST_FROM_ISR 0
+
+bool etcpal_signal_create(etcpal_signal_t* id);
+bool etcpal_signal_wait(etcpal_signal_t* id);
+bool etcpal_signal_try_wait(etcpal_signal_t* id);
+bool etcpal_signal_timed_wait(etcpal_signal_t* id, int timeout_ms);
+void etcpal_signal_post(etcpal_signal_t* id);
+#define etcpal_signal_post_from_isr etcpal_signal_post
+void etcpal_signal_destroy(etcpal_signal_t* id);
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif /* ETCPAL_OS_SIGNAL_H_ */
